Rename files in dsl-text
