+++ .changeset/empty-geckos-try.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-extension-dsl-text': major
+---
+
+Rename files prefix from `DSLText` to `DSL_Text`.
