+++ packages/legend-extension-dsl-text/src/components/studio/DSL_Text_CodeSnippets.ts
