+++ packages/legend-extension-dsl-text/src/index.ts
@@ -14,5 +14,5 @@
  * limitations under the License.
  */
 
-export * from './DSLText_Extension.js';
-export { DSLText_LegendStudioApplicationPlugin } from './components/studio/DSLText_LegendStudioApplicationPlugin.js';
+export * from './DSL_Text_Extension.js';
+export { DSL_Text_LegendStudioApplicationPlugin } from './components/studio/DSL_Text_LegendStudioApplicationPlugin.js';
