+++ packages/legend-extension-dsl-text/src/stores/studio/DSL_Text_GraphModifierHelper.ts
@@ -15,7 +15,7 @@
  */
 
 import { action } from 'mobx';
-import type { Text } from '../../graph/metamodel/pure/model/packageableElements/text/DSLText_Text.js';
+import type { Text } from '../../graph/metamodel/pure/model/packageableElements/text/DSL_Text_Text.js';
 
 export const text_setType = action((text: Text, type: string): void => {
   text.type = type;
