+++ packages/legend-extension-dsl-text/src/stores/studio/DSL_Text_LegendStudioApplicationNavigationContext.ts
